backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
SQLAlchemy==2.0.23
asyncpg
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
greenlet==3.0.1

python-jose[cryptography]==3.3.0
argon2-cffi
python-multipart
